Strømlinjeforming av integrasjonstester
* Strømlinjeformet oppsettet av tester som bruker kafka-embeddet.
* Fjernet siste rest av assertions fra Junit5, og bruker heller Kluent.

PB-201. Integrasjonstester i isolasjon
