+++ src/intTest/kotlin/no/nav/personbruker/dittnav/eventaggregator/kafka/StringBasicKafkaEmbeddedTesting.kt
@@ -5,10 +5,10 @@ import no.nav.common.JAASCredential
 import no.nav.common.KafkaEnvironment
 import no.nav.personbruker.dittnav.eventaggregator.util.KafkaConsumerUtil
 import no.nav.personbruker.dittnav.eventaggregator.util.KafkaProducerUtil
+import org.amshove.kluent.`should equal`
 import org.amshove.kluent.shouldContainAll
 import org.amshove.kluent.shouldEqualTo
 import org.junit.jupiter.api.AfterAll
-import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
 
 /**
@@ -43,7 +43,7 @@ class StringBasicKafkaEmbeddedTesting {
 
     @Test
     fun `Kafka-instansen i minnet har blitt staret`() {
-        Assertions.assertEquals(embeddedEnv.serverPark.status, KafkaEnvironment.ServerParkStatus.Started)
+        embeddedEnv.serverPark.status `should equal` KafkaEnvironment.ServerParkStatus.Started
     }
 
     @Test
