+++ src/intTest/kotlin/no/nav/personbruker/dittnav/eventaggregator/util/KafkaTestUtil.kt
@@ -1,0 +1,41 @@
+package no.nav.personbruker.dittnav.eventaggregator.util
+
+import no.nav.common.JAASCredential
+import no.nav.common.KafkaEnvironment
+import no.nav.personbruker.dittnav.eventaggregator.config.Environment
+import org.apache.avro.generic.GenericRecord
+
+object KafkaTestUtil {
+
+    val username = "srvkafkaclient"
+    val password = "kafkaclient"
+
+    fun createDefaultKafkaEmbeddedInstance(topics: List<String>): KafkaEnvironment {
+        return KafkaEnvironment(
+                topicNames = topics,
+                withSecurity = true,
+                withSchemaRegistry = true,
+                users = listOf(JAASCredential(username, password))
+        )
+    }
+
+    fun createEnvironmentForEmbeddedKafka(embeddedEnv: KafkaEnvironment): Environment {
+        return Environment().copy(
+                bootstrapServers = embeddedEnv.brokersURL.substringAfterLast("/"),
+                schemaRegistryUrl = embeddedEnv.schemaRegistry!!.url,
+                username = username,
+                password = password
+        )
+    }
+
+    suspend fun produceEvents(env: Environment, topicName: String, events: Map<String, GenericRecord>): Boolean {
+        return KafkaProducerUtil.kafkaAvroProduce(
+                env.bootstrapServers,
+                env.schemaRegistryUrl,
+                topicName,
+                env.username,
+                env.password,
+                events)
+    }
+
+}
